move queries left right to prepare as they are shared between queries
