+++ src/fibonacci/quotients.rs
@@ -1,7 +1,8 @@
 use crate::constraints::DenominatorInverseHint;
 use crate::fibonacci::fiat_shamir::FSOutput;
 use crate::fibonacci::prepare::PrepareOutput;
-use crate::fibonacci::PerQueryQuotientHint;
+use crate::precomputed_merkle_tree::PrecomputedMerkleTreeProof;
+use crate::treepp::pushable::{Builder, Pushable};
 use stwo_prover::core::fft::ibutterfly;
 use stwo_prover::core::fields::qm31::QM31;
 use stwo_prover::core::fields::FieldExpOps;
@@ -43,49 +44,27 @@ pub fn compute_quotients_hints(
             ),
         ];
 
-        let mut queried_values_left = vec![];
-        let mut queried_values_right = vec![];
-        for (trace, composition) in prepare_output
-            .merkle_proofs_traces
-            .iter()
-            .zip(prepare_output.merkle_proofs_compositions.iter())
-        {
-            let mut left_vec = vec![];
-            let mut right_vec = vec![];
-            for (&left, &right) in trace
-                .left
-                .iter()
-                .zip(trace.right.iter())
-                .chain(composition.left.iter().zip(composition.right.iter()))
-            {
-                left_vec.push(left);
-                right_vec.push(right);
-            }
-            queried_values_left.push(left_vec);
-            queried_values_right.push(right_vec);
-        }
-
         let mut nominators = vec![];
         for column_line_coeff in prepare_output.column_line_coeffs.iter().take(3) {
             nominators.push(column_line_coeff.apply_twin(
                 precomputed.circle_point,
-                &[queried_values_left[i][0]],
-                &[queried_values_right[i][0]],
+                &[prepare_output.queried_values_left[i][0]],
+                &[prepare_output.queried_values_right[i][0]],
             ));
         }
         nominators.push(prepare_output.column_line_coeffs[3].apply_twin(
             precomputed.circle_point,
             &[
-                queried_values_left[i][1],
-                queried_values_left[i][2],
-                queried_values_left[i][3],
-                queried_values_left[i][4],
+                prepare_output.queried_values_left[i][1],
+                prepare_output.queried_values_left[i][2],
+                prepare_output.queried_values_left[i][3],
+                prepare_output.queried_values_left[i][4],
             ],
             &[
-                queried_values_right[i][1],
-                queried_values_right[i][2],
-                queried_values_right[i][3],
-                queried_values_right[i][4],
+                prepare_output.queried_values_right[i][1],
+                prepare_output.queried_values_right[i][2],
+                prepare_output.queried_values_right[i][3],
+                prepare_output.queried_values_right[i][4],
             ],
         ));
 
@@ -141,4 +120,38 @@ pub fn compute_quotients_hints(
     }
 
     (QuotientsOutput { fold_results }, hints)
+}
+
+#[derive(Default, Clone)]
+/// Hint that repeats for each query.
+pub struct PerQueryQuotientHint {
+    /// Precomputed tree Merkle proofs.
+    pub precomputed_merkle_proofs: Vec<PrecomputedMerkleTreeProof>,
+
+    /// Denominator inverse hints.
+    pub denominator_inverse_hints: Vec<DenominatorInverseHint>,
+
+    /// Test-only: the FRI answer.
+    pub test_only_fri_answer: Vec<QM31>,
+}
+
+impl Pushable for &PerQueryQuotientHint {
+    fn bitcoin_script_push(self, mut builder: Builder) -> Builder {
+        for proof in self.precomputed_merkle_proofs.iter() {
+            builder = proof.bitcoin_script_push(builder);
+        }
+        for hint in self.denominator_inverse_hints.iter() {
+            builder = hint.bitcoin_script_push(builder);
+        }
+        for elem in self.test_only_fri_answer.iter().rev() {
+            builder = elem.bitcoin_script_push(builder);
+        }
+        builder
+    }
+}
+
+impl Pushable for PerQueryQuotientHint {
+    fn bitcoin_script_push(self, builder: Builder) -> Builder {
+        (&self).bitcoin_script_push(builder)
+    }
 }
